+++ LinearAlgebraAssessment/clock.h
@@ -6,11 +6,11 @@ using delta_time = timer::duration;
 
 inline timer::time_point now() { return timer::now(); }
 
-inline float to_seconds(delta_time dt) {
-	return std::chrono::duration_cast<std::chrono::duration<float>>(dt).count();
+inline double to_seconds(delta_time dt) {
+	return std::chrono::duration_cast<std::chrono::duration<double>>(dt).count();
 }
 
-inline delta_time from_seconds(float sec) {
+inline delta_time from_seconds(double sec) {
 	return std::chrono::duration_cast<delta_time>(
-		std::chrono::duration<float>(sec));
+		std::chrono::duration<double>(sec));
 }
